% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_data.R
\name{inverse_scale}
\alias{inverse_scale}
\title{Inverse-Transform Scaled Data}
\usage{
inverse_scale(scaled_data, scale_result)
}
\arguments{
\item{scaled_data}{A \code{data.frame} of scaled values.}

\item{scale_result}{The list returned by \code{\link[=scale_data]{scale_data()}}.}
}
\value{
A \code{data.frame} with the scaled columns back in their original units.
}
\description{
Reverses the scaling applied by \code{\link[=scale_data]{scale_data()}} to recover original units.
}
\examples{
data(airquality)
clean <- airquality[complete.cases(airquality), ]
sp    <- split_data(clean, verbose = FALSE)
sc    <- scale_data(sp, verbose = FALSE)
original_train <- inverse_scale(sc$train_scaled, sc)

}
